Use Controller for inventory endpoint
Inventory for cli has been moved to tool/inventory.php

Use constant instead of hardcoded string
