+++ tools/inventory.php
@@ -32,37 +32,30 @@
  * ---------------------------------------------------------------------
  */
 
-namespace Glpi\Controller;
+use Glpi\Inventory\Conf;
+use Glpi\Inventory\Request;
 
-use Glpi\Api\APIRest;
-use Glpi\Application\ErrorHandler;
-use Glpi\Http\Firewall;
-use Glpi\Http\HeaderlessStreamedResponse;
-use Glpi\Security\Attribute\SecurityStrategy;
-use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\Routing\Attribute\Route;
-
-final class ApiRestController extends AbstractController
-{
-    #[Route(
-        "/apirest.php{request_parameters}",
-        name: "glpi_api_rest",
-        requirements: [
-            'request_parameters' => '.*',
-        ]
-    )]
-    #[SecurityStrategy(Firewall::STRATEGY_NO_CHECK)]
-    public function __invoke(Request $request): Response
-    {
-        $_SERVER['PATH_INFO'] = $request->get('request_parameters');
-
-        return new HeaderlessStreamedResponse(function () {
-            // Ensure errors will not break API output.
-            ErrorHandler::getInstance()->disableOutput();
-
-            $api = new APIRest();
-            $api->call();
-        });
-    }
+$conf = new Conf();
+if ($conf->enabled_inventory != 1) {
+    die("Inventory is disabled");
 }
+
+if (!isCommandLine()) {
+    die('This script is only available from the command line');
+}
+
+$f = fopen('php://stdin', 'r');
+$contents = '';
+while ($line = fgets($f)) {
+    $contents .= $line;
+}
+fclose($f);
+
+try {
+    $inventory_request = new Request();
+    $inventory_request->handleRequest($contents);
+} catch (\Throwable $e) {
+    $inventory_request->addError($e->getMessage());
+}
+
+$inventory_request->handleMessages();
